move aug_comparator.py and aug_out_filter.py to utils
